# ecommerce-app/openapi.yaml
openapi: 3.1.0
...
info:
  title: ShopKart E-Commerce Storefront API
  version: 1.0.0
  summary: Customer-facing storefront API for the ShopKart e-commerce application.
  description: |
    REST API powering the ShopKart storefront: authentication, profile & address
    management, product catalogue browsing, cart & coupon handling, checkout /
    orders, and verified product reviews.

    ## Authentication
    Stateless **JWT Bearer** tokens. Obtain a token from `POST /api/auth/signup`
    or `POST /api/auth/login`, then send it as `Authorization: Bearer <token>`.
    Tokens expire after 7 days (configurable via `JWT_EXPIRES_IN`). Logout is
    client-side (drop the token); `POST /api/auth/logout` is a no-op convenience.

    ## Core domain rule
    A purchase is recorded immutably as `user -> order -> order_items`. Each
    `order_item` snapshots the product name, variant name, SKU and unit price
    paid at that time, so order history never changes if a product is later
    renamed, repriced or archived.

    ## Pricing
    - Tax: 8% of `(subtotal - discount)`.
    - Shipping: flat ₹49, waived when `subtotal >= ₹999` (or cart is empty).
    - Coupons: `percent` (optionally capped via `max_discount`) or `fixed`.

    ## Conventions
    - All request and response bodies are `application/json` unless noted.
    - All monetary values are in INR (₹).
    - Timestamps are RFC 3339 / ISO 8601 date-time in UTC.
  contact:
    name: ShopKart API Support
...
  license:
    name: MIT
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /api/products:
    get:
      tags: [Products]
      summary: Browse, search, filter and sort products
      description: |
        Returns active (non-archived) products, each decorated with price range,
        aggregate rating, total stock across variants, and a primary image.
        Supports full-text-ish search (`q`), category and price/rating/stock
        filters, sorting and simple page-based pagination.
      operationId: listProducts
      security: []
      parameters:
        - $ref: '#/components/parameters/SearchQuery'
        - $ref: '#/components/parameters/CategoryFilter'
        - $ref: '#/components/parameters/MinPrice'
        - $ref: '#/components/parameters/MaxPrice'
        - $ref: '#/components/parameters/MinRating'
        - $ref: '#/components/parameters/InStock'
        - $ref: '#/components/parameters/Sort'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: A page of products.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProductListResponse' }
              example:
                products:
                  - id: 1
                    name: Aurora Wireless Headphones
                    slug: aurora-wireless-headphones
                    category: Electronics
                    brand: Aurora
                    description: Over-ear ANC headphones with 40h battery.
                    is_archived: false
                    created_at: '2026-07-01T09:00:00Z'
                    min_price: '4999.00'
                    max_price: '5999.00'
                    total_stock: '37'
                    avg_rating: '4.60'
                    review_count: '18'
                    image: https://cdn.example.com/products/aurora-1.jpg
                page: 1
                limit: 24
...
  /api/products/{idOrSlug}:
    get:
      tags: [Products]
      summary: Get a single product by numeric id or slug
      description: |
        Accepts either a numeric product `id` or a URL `slug`. Returns the
        product with its images, active variants, reviews and aggregate rating.
      operationId: getProduct
      security: []
      parameters:
        - name: idOrSlug
          in: path
          required: true
          description: Numeric product id (e.g. `12`) or product slug (e.g. `aurora-wireless-headphones`).
          schema:
            type: string
          examples:
            byId: { value: '12', summary: By numeric id }
            bySlug: { value: aurora-wireless-headphones, summary: By slug }
      responses:
        '200':
          description: Product detail.
          content:
            application/json:
              schema:
                type: object
                required: [product]
                properties:
                  product: { $ref: '#/components/schemas/ProductDetail' }
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/orders:
    get:
      tags: [Orders]
      summary: List my orders
      description: Order history for the authenticated user, newest first.
      operationId: listOrders
      responses:
        '200':
          description: Order history.
          content:
            application/json:
              schema:
                type: object
                required: [orders]
                properties:
                  orders:
                    type: array
                    items: { $ref: '#/components/schemas/OrderListItem' }
              example:
                orders:
                  - id: 1001
                    order_number: SK-2026-001001
                    status: paid
                    payment_status: paid
                    total: '9718.06'
                    placed_at: '2026-07-24T10:35:00Z'
                    item_count: 2
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Orders]
      summary: Place an order from the cart
      description: |
        Creates an order from the current cart. Re-checks stock (locking variant
        rows), snapshots the chosen delivery address and each line item, applies
        an optional coupon, decrements stock and clears the cart — all in one
        transaction. The order starts as `status: placed`, `payment_status: unpaid`.
      operationId: createOrder
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/OrderCreateRequest' }
            example: { address_id: 15, coupon_code: WELCOME10 }
      responses:
        '201':
          description: Order placed. Returns the new order id and number.
          content:
            application/json:
              schema:
                type: object
                required: [order]
                properties:
                  order: { $ref: '#/components/schemas/OrderRef' }
              example:
                order: { id: 1001, order_number: SK-2026-001001 }
        '400':
          description: Missing address, empty cart, or invalid coupon.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              examples:
                address: { value: { error: 'Please choose a delivery address.' } }
                empty: { value: { error: 'Your cart is empty.' } }
                coupon: { value: { error: 'Invalid coupon code.' } }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: The delivery address was not found for this user.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example: { error: 'Delivery address not found.' }
        '409':
          description: A cart item became unavailable or short on stock during checkout.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              examples:
                unavailable: { value: { error: '"Aurora Wireless Headphones" is no longer available.' } }
                stock: { value: { error: 'Only 1 of "Aurora Wireless Headphones" left in stock.' } }
  /api/orders/{id}:
    parameters:
      - $ref: '#/components/parameters/OrderId'
    get:
      tags: [Orders]
      summary: Get a single order with its line items
      operationId: getOrder
      responses:
        '200':
          description: The order and its immutable line items.
          content:
            application/json:
              schema:
                type: object
                required: [order]
                properties:
                  order: { $ref: '#/components/schemas/Order' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT Bearer token obtained from `/api/auth/login` or `/api/auth/signup`.
        Send as `Authorization: Bearer <token>`. Payload carries `id`, `email`,
        `role` and `name`; expires after 7 days by default.

  # ─────────────────────────────  PARAMETERS  ─────────────────────────────
  parameters:
...
    OrderId:
      name: id
      in: path
      required: true
      description: Order identifier owned by the authenticated user.
      schema: { type: integer, format: int32, minimum: 1 }
      example: 1001
...
    SearchQuery:
      name: q
      in: query
      required: false
      description: Case-insensitive keyword matched against product name, description and brand.
      schema: { type: string }
      example: headphones
    CategoryFilter:
      name: category
      in: query
      required: false
      description: Exact category name to filter by (see `/api/products/categories`).
      schema: { type: string }
      example: Electronics
    MinPrice:
      name: min_price
      in: query
      required: false
      description: Minimum product price (INR), inclusive.
      schema: { type: number, minimum: 0 }
      example: 1000
    MaxPrice:
      name: max_price
      in: query
      required: false
      description: Maximum product price (INR), inclusive.
      schema: { type: number, minimum: 0 }
      example: 6000
    MinRating:
      name: min_rating
      in: query
      required: false
      description: Minimum average rating, inclusive.
      schema: { type: number, minimum: 0, maximum: 5 }
      example: 4
    InStock:
      name: in_stock
      in: query
      required: false
      description: When `true`, only products with total stock greater than zero are returned.
      schema: { type: string, enum: ['true', 'false'] }
      example: 'true'
    Sort:
      name: sort
      in: query
      required: false
      description: Sort order. Defaults to `newest` when omitted or unrecognised.
      schema:
        type: string
        enum: [price_asc, price_desc, newest, rating]
        default: newest
      example: rating
    Page:
      name: page
      in: query
      required: false
      description: 1-based page number.
      schema: { type: integer, format: int32, minimum: 1, default: 1 }
      example: 1
    Limit:
      name: limit
      in: query
      required: false
      description: Page size. Clamped server-side to the range 1–100.
      schema: { type: integer, format: int32, minimum: 1, maximum: 100, default: 24 }
      example: 24

  # ─────────────────────────────  RESPONSES  ─────────────────────────────
  responses:
...
    Unauthorized:
      description: Authentication is required or the token is invalid/expired.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          examples:
            missing: { value: { error: 'Please log in to continue.' } }
            expired: { value: { error: 'Your session has expired. Please log in again.' } }
    NotFound:
      description: The requested resource does not exist (or is not owned by the caller).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: 'Not found.' }
...
  schemas:
    Error:
      type: object
      description: |
        Standard error body returned by every failing endpoint. The API returns
        a single human-readable `error` message (no code/details envelope).
      required: [error]
      additionalProperties: false
      properties:
        error:
          type: string
          description: Human-readable, client-safe error message.
          example: 'Please log in to continue.'
...
    DecimalString:
      type: string
      description: |
        A fixed-precision decimal serialised as a string (PostgreSQL NUMERIC).
        Parse before arithmetic. Two decimal places for currency.
      pattern: '^-?\d+(\.\d+)?$'
      example: '4999.00'

    # ── Users & auth ──────────────────────────────────────────────────────
...
    ProductListItem:
      type: object
      description: A product decorated with aggregate price/rating/stock and a primary image.
      required: [id, name, category, is_archived, min_price, max_price, total_stock, avg_rating, review_count]
      properties:
        id: { type: integer, format: int32, example: 1 }
        name: { type: string, example: Aurora Wireless Headphones }
        slug: { type: [string, 'null'], example: aurora-wireless-headphones }
        category: { type: string, example: Electronics }
        brand: { type: [string, 'null'], example: Aurora }
        description: { type: [string, 'null'], example: Over-ear ANC headphones with 40h battery. }
        is_archived: { type: boolean, example: false }
        created_at: { type: string, format: date-time, example: '2026-07-01T09:00:00Z' }
        min_price: { $ref: '#/components/schemas/DecimalString' }
        max_price: { $ref: '#/components/schemas/DecimalString' }
        total_stock:
          type: string
          description: Sum of stock across active variants (NUMERIC/bigint as string).
          example: '37'
        avg_rating: { $ref: '#/components/schemas/DecimalString' }
        review_count:
          type: string
          description: Number of reviews (bigint serialised as string).
          example: '18'
        image:
          type: [string, 'null']
          format: uri
          description: URL of the primary product image, or null if none.
          example: https://cdn.example.com/products/aurora-1.jpg
    ProductImage:
      type: object
      required: [id, url, position]
      properties:
        id: { type: integer, format: int32, example: 501 }
        url: { type: string, format: uri, example: https://cdn.example.com/products/aurora-1.jpg }
        position: { type: integer, format: int32, description: Display order (ascending)., example: 0 }
    ProductVariant:
      type: object
      description: A purchasable SKU of a product.
      required: [id, sku, variant_name, attributes, price, stock, is_active]
      properties:
        id: { type: integer, format: int32, example: 108 }
        sku: { type: string, example: AUR-HP-BLK }
        variant_name: { type: string, default: Default, example: Black }
        attributes:
          type: object
          description: Free-form key/value attributes (e.g. size, colour).
          additionalProperties: true
          example: { color: Black, edition: Standard }
        price: { $ref: '#/components/schemas/DecimalString' }
        stock: { type: integer, format: int32, minimum: 0, example: 12 }
        is_active: { type: boolean, example: true }
    Review:
      type: object
      description: Public-facing review (author name only).
      required: [id, rating, created_at, author]
      properties:
        id: { type: integer, format: int32, example: 5 }
        rating: { type: integer, minimum: 1, maximum: 5, example: 5 }
        title: { type: [string, 'null'], example: Fantastic sound }
        body: { type: [string, 'null'], example: Battery really does last all week. }
        created_at: { type: string, format: date-time, example: '2026-07-20T14:12:00Z' }
        updated_at: { type: string, format: date-time, example: '2026-07-20T14:12:00Z' }
        author: { type: string, description: Display name of the reviewer., example: Aarav Sharma }
...
    ProductDetail:
      type: object
      description: Full product with images, active variants, reviews and aggregates.
      required: [id, name, category, base_price, is_archived, images, variants, reviews, avg_rating, review_count]
      properties:
        id: { type: integer, format: int32, example: 1 }
        name: { type: string, example: Aurora Wireless Headphones }
        slug: { type: [string, 'null'], example: aurora-wireless-headphones }
        description: { type: [string, 'null'], example: Over-ear ANC headphones with 40h battery. }
        category: { type: string, example: Electronics }
        brand: { type: [string, 'null'], example: Aurora }
        base_price: { $ref: '#/components/schemas/DecimalString' }
        is_archived: { type: boolean, example: false }
        created_at: { type: string, format: date-time, example: '2026-07-01T09:00:00Z' }
        images:
          type: array
          items: { $ref: '#/components/schemas/ProductImage' }
        variants:
          type: array
          items: { $ref: '#/components/schemas/ProductVariant' }
        reviews:
          type: array
          items: { $ref: '#/components/schemas/Review' }
        avg_rating: { $ref: '#/components/schemas/DecimalString' }
        review_count: { type: integer, format: int32, example: 18 }
    ProductListResponse:
      type: object
      required: [products, page, limit]
      properties:
        products:
          type: array
          items: { $ref: '#/components/schemas/ProductListItem' }
        page: { type: integer, format: int32, description: Echoed current page., example: 1 }
        limit: { type: integer, format: int32, description: Effective (clamped) page size., example: 24 }

    # ── Cart ──────────────────────────────────────────────────────────────
...
    OrderCreateRequest:
      type: object
      required: [address_id]
      additionalProperties: false
      properties:
        address_id:
          type: integer
          format: int32
          minimum: 1
          description: Id of a delivery address owned by the user.
          example: 15
        coupon_code:
          type: [string, 'null']
          description: Optional coupon code to apply to the order.
          example: WELCOME10
    OrderRef:
      type: object
      description: Minimal reference returned when an order is placed.
      required: [id, order_number]
      properties:
        id: { type: integer, format: int32, example: 1001 }
        order_number: { type: string, example: SK-2026-001001 }
    OrderListItem:
      type: object
      required: [id, order_number, status, payment_status, total, placed_at, item_count]
      properties:
        id: { type: integer, format: int32, example: 1001 }
        order_number: { type: string, example: SK-2026-001001 }
        status: { $ref: '#/components/schemas/OrderStatus' }
        payment_status: { $ref: '#/components/schemas/PaymentStatus' }
        total: { $ref: '#/components/schemas/DecimalString' }
        placed_at: { type: string, format: date-time, example: '2026-07-24T10:35:00Z' }
        item_count: { type: integer, format: int32, description: Total quantity across line items., example: 2 }
    OrderStatus:
      type: string
      description: Lifecycle status of an order.
      enum: [placed, paid, packed, shipped, delivered, cancelled]
      example: paid
    PaymentStatus:
      type: string
      enum: [unpaid, paid]
      example: paid
    OrderItem:
      type: object
      description: Immutable snapshot of a purchased line item.
      required: [id, product_name, variant_name, sku, unit_price, quantity, line_total]
      properties:
        id: { type: integer, format: int32, example: 900 }
        product_id:
          type: [integer, 'null']
          format: int32
          description: Null if the product was later deleted.
          example: 1
        variant_id:
          type: [integer, 'null']
          format: int32
          description: Null if the variant was later deleted.
          example: 108
        product_name: { type: string, example: Aurora Wireless Headphones }
        variant_name: { type: string, example: Black }
        sku: { type: string, example: AUR-HP-BLK }
        unit_price: { $ref: '#/components/schemas/DecimalString' }
        quantity: { type: integer, format: int32, minimum: 1, example: 2 }
        line_total: { $ref: '#/components/schemas/DecimalString' }
    Order:
      type: object
      description: Full order with snapshotted address, totals, tracking and line items.
      required: [id, order_number, user_id, status, payment_status, subtotal, discount, tax, shipping_fee, total, placed_at, items]
      properties:
        id: { type: integer, format: int32, example: 1001 }
        order_number: { type: string, example: SK-2026-001001 }
        user_id: { type: integer, format: int32, example: 42 }
        status: { $ref: '#/components/schemas/OrderStatus' }
        payment_status: { $ref: '#/components/schemas/PaymentStatus' }
        subtotal: { $ref: '#/components/schemas/DecimalString' }
        discount: { $ref: '#/components/schemas/DecimalString' }
        tax: { $ref: '#/components/schemas/DecimalString' }
        shipping_fee: { $ref: '#/components/schemas/DecimalString' }
        total: { $ref: '#/components/schemas/DecimalString' }
        coupon_code: { type: [string, 'null'], example: WELCOME10 }
        ship_recipient: { type: [string, 'null'], example: Aarav Sharma }
        ship_phone: { type: [string, 'null'], example: '+919812345678' }
        ship_line1: { type: [string, 'null'], example: 12 Residency Road }
        ship_line2: { type: [string, 'null'], example: Apt 4B }
        ship_city: { type: [string, 'null'], example: Bengaluru }
        ship_state: { type: [string, 'null'], example: Karnataka }
        ship_postal: { type: [string, 'null'], example: '560025' }
        ship_country: { type: [string, 'null'], example: India }
        tracking_carrier: { type: [string, 'null'], example: BlueDart }
        tracking_number: { type: [string, 'null'], example: BD123456789IN }
        placed_at: { type: string, format: date-time, example: '2026-07-24T10:35:00Z' }
        paid_at: { type: [string, 'null'], format: date-time, example: '2026-07-24T10:36:10Z' }
        packed_at: { type: [string, 'null'], format: date-time, example: '2026-07-24T15:00:00Z' }
        shipped_at: { type: [string, 'null'], format: date-time, example: '2026-07-25T09:20:00Z' }
        delivered_at: { type: [string, 'null'], format: date-time, example: null }
        cancelled_at: { type: [string, 'null'], format: date-time, example: null }
        items:
          type: array
          items: { $ref: '#/components/schemas/OrderItem' }
...